moving over to parallel file
